refactor: refactor outgoingRequestList(TNA-92)
